ISO_15924 {ISOcodes}	R Documentation

ISO 15924 Script Codes

Description

International Organization for Standardization (ISO) codes for the representation of names of scripts (writing systems, “a set of graphic characters used for the written form of one or more languages”).

Usage

ISO_15924

Format

ISO_15924 is a data frame with variables

Alpha_4

a character vector with the 4-letter (alpha-4) ISO 15924 script codes.

Numeric

a character vector with numeric script codes providing some measure of mnemonicity.

Name

a character vector with the (English) script names.

PVA

a character vector with the Property Value Alias defined by Unicode (ISO 10646) (if available).

Date

a Date object with the date the script was registered.

Details

The following number ranges are used for the numeric codes:

000-099 	 Hieroglyphic and cuneiform scripts

100-199 	 Right-to-left alphabetic scripts

200-299 	 Left-to-right alphabetic scripts

300-399 	 Alphasyllabic scripts

400-499 	 Syllabic scripts

500-599 	 Ideographic scripts

600-699 	 Undeciphered scripts

700-899 	 (unassigned)

900-999 	 Private use, aliases, special codes

Source

http://www.unicode.org/iso15924/

References

http://en.wikipedia.org/wiki/ISO_15924

[Package ISOcodes version 2020.03.16 Index]