ISO_3166 {ISOcodes}	R Documentation

ISO 3166 Country Codes

Description

International Organization for Standardization (ISO) codes for the representation of names of countries and their subdivisions. Consists of three parts. Part 1: Country codes, defines codes for country and dependent area names. Part 2: Country subdivision code, defines codes for the principal subdivisions of a country or dependent area. Part 3: Code for formerly used names of countries, defines codes for superseded ISO 3166-1 codes.

Usage

ISO_3166_1
ISO_3166_2
ISO_3166_3

Format

ISO_3166_1 is a character frame with variables Alpha_2, Alpha_3, and Numeric (giving the two-letter, three-letter and three-digit numeric country codes) and Name, Official_name, and Common_name (giving the respective names).

ISO_3166_2 is a character frame with variables Code, Type, Name, and Parent, giving the code, type and name of the subdivision, and a parent subdivision in case this is different from the country.

ISO_3166_3 is a character frame with variables Alpha_4 (the 4-letter code of the retired country), Alpha_3, Numeric, and Name (the original 3166-1 code elements of the country), and Date_withdrawn and Comment.

Source

Converted from JSON files provided by Debian's iso-codes package (https://salsa.debian.org/iso-codes-team/iso-codes).

References

http://en.wikipedia.org/wiki/ISO_3166

[Package ISOcodes version 2019.04.22 Index]