ISO_4217 {ISOcodes}	R Documentation

ISO 4217 Currency Codes

Description

International Organization for Standardization (ISO) codes for the representation of currencies.

Usage

ISO_4217
ISO_4217_Historic

Format

ISO_4217 is a character frame with variables Letter, Numeric and Currency, giving the 3-letter and 3-digit codes and the names of the respective currency.

ISO_4217_Historic is a character frame with the currency codes retired from ISO 4217, containing variable Date_withdrawn in addition to the variables in ISO_4217.

Source

Converted from XML files provided by Debian's iso-codes package (https://salsa.debian.org/iso-codes-team/iso-codes).

References

http://en.wikipedia.org/wiki/ISO_4217

[Package ISOcodes version 2018.06.29 Index]