ISO_8859 {ISOcodes}	R Documentation

ISO 8859 Character Codes

Description

International Organization for Standardization (ISO) codes for 8-bit character encodings for use by computers. The data set gives the maps of the characters to Unicode (i.e., the respective ISO 10646 codes).

Usage

ISO_8859

Format

A character array of dimension 256 \times 15 \times 3, with the first dimension corresponding to the character codes from 0 to 255 (0x00 to 0xff), the second to the parts of the ISO 8859 standard, and the third to the Unicode (ISO 10646) code and name, and the respective character.

Details

The ISO 8859, more formally ISO/IEC 8859, standard is divided into numbered, separately published parts, such as as ISO/IEC 8859-1, ISO/IEC 8859-2, etc., each of which may be informally referred to as a standard in itself. There are currently 15 parts as of 2006 excluding the abandoned ISO/IEC 8859-12 standard:

Part 1 	 Latin-1 Western European

Part 2 	 Latin-2 Central European

Part 3 	 Latin-3 South European

Part 4 	 Latin-4 North European

Part 5 	 Latin/Cyrillic

Part 6 	 Latin/Arabic

Part 7 	 Latin/Greek

Part 8 	 Latin/Hebrew

Part 9 	 Latin-5 Turkish

Part 10 	 Latin-6 Nordic

Part 11 	 Latin/Thai

Part 13 	 Latin-7 Baltic Rim

Part 14 	 Latin-8 Celtic

Part 15 	 Latin-9

Part 16 	 Latin-10

Source

http://unicode.org/Public/MAPPINGS/ISO8859.

References

http://en.wikipedia.org/wiki/ISO_8859

Examples

## ISO 8859 characters at position 200 (number 199).
data("ISO_8859")
ISO_8859[200, , ]

[Package ISOcodes version 2019.04.22 Index]