UN_M.49 {ISOcodes}	R Documentation

UN M.49 Area Codes

Description

Country and area code classifications (M49) from the United Nations Statistics Division.

Usage

UN_M.49_Countries
UN_M.49_Regions

Details

UN M.49 is a standard for area codes used by the United Nations for statistical purposes. Each area code is a 3-digit number which can refer to a wide variety of geographical, political, or economic regions, like a continent, a country, or a specific group of developing countries.

UN_M.49_Countries contains the codes for countries and areas as a character frame with variables Code, Name and ISO_Alpha_3 giving the 3-letter UN M.49 code and name and the respective alpha-3 ISO 3166 code.

UN_M.49_Regions contains the codes for the composition of macro geographical (continental) regions, geographical sub-regions, and selected economic and other groupings as a data frame with the character variables Code, Name, Parent and Children giving the 3-letter UN M.49 code, name, the code of the parent area and the codes of children areas separated by , , respectively, and variable Type, a factor with levels "Region" or "Grouping".

Source

http://unstats.un.org/unsd/methods/m49/m49.htm

Examples

## Name and codes of countries in Southern Europe:
data("UN_M.49_Regions")
data("UN_M.49_Countries")
region <- subset(UN_M.49_Regions, Name == "Southern Europe")
codes <- unlist(strsplit(region$Children, ", "))
subset(UN_M.49_Countries, Code %in% codes)

[Package ISOcodes version 2019.04.22 Index]